refactor(examples): Restructure RMRK example so it reflects program concept
