+++ examples/rmrk/catalog/app/Cargo.toml
@@ -6,6 +6,6 @@ edition = "2021"
 [dependencies]
 parity-scale-codec = { workspace = true, features = ["derive"] }
 sails-macros.workspace = true
-sails-rtl.workspace = true
+sails-rtl-gstd.workspace = true
 sails-service-meta.workspace = true
 scale-info = { workspace = true, features = ["derive"] }
