+++ examples/rmrk/catalog/app/src/lib.rs
@@ -1,219 +1,62 @@
 #![no_std]
 
-use errors::Error;
-use parts::{CollectionId, Part, PartId, SlotPart};
-use sails_macros::gservice;
-use sails_rtl::{
-    collections::{BTreeMap, BTreeSet},
-    Result as RtlResult, *,
-};
+use sails_rtl_gstd::GStdExecContext;
+use services::Catalog;
 
-pub mod errors;
-pub mod parts;
-
-static mut CATALOG_ADMIN: Option<ActorId> = None;
-
-type Result<T> = RtlResult<T, Error>;
-
-type PartMap<K, V> = BTreeMap<K, V>;
+// Exposed publicly because of service metadata for IDL
+// Will be superseded by program metadata, thus pub won't be needed
+pub mod services;
 
 #[derive(Default)]
-pub struct CatalogData {
-    parts: PartMap<PartId, Part>,
-    is_equippable_to_all: BTreeSet<PartId>,
+pub struct Program;
+
+impl Program {
+    // Initialize program and seed hosted services
+    pub fn new() -> Self {
+        let exec_context = GStdExecContext::default();
+        Catalog::seed(exec_context);
+        Self
+    }
+
+    // Expose hosted service
+    pub fn catalog(&self) -> Catalog<GStdExecContext> {
+        let exec_context = GStdExecContext::default();
+        Catalog::new(exec_context)
+    }
+
+    // Will be generated by the `ginject_defaults` macro from the `new` method
+    fn __new(exec_context: GStdExecContext) -> Self {
+        Catalog::seed(exec_context);
+        Self
+    }
+
+    // Will be generated by the `ginject_defaults` macro from the `catalog` method
+    fn __catalog(&self, exec_context: GStdExecContext) -> Catalog<GStdExecContext> {
+        Catalog::new(exec_context)
+    }
 }
 
-pub struct Catalog<'a, TExecContext: ExecContext> {
-    data: &'a mut CatalogData,
-    exec_context: TExecContext,
-}
+// The below code will be generated by the `gprogram` macro
+#[cfg(target_arch = "wasm32")]
+pub mod wasm {
+    use super::*;
+    use sails_rtl_gstd::{gstd, gstd::msg};
+    use services::requests;
 
-impl<'a, TExecContext: ExecContext> Catalog<'a, TExecContext>
-where
-    TExecContext: ExecContext,
-{
-    pub fn new(data: &'a mut CatalogData, exec_context: TExecContext) -> Self {
+    static mut PROGRAM: Option<Program> = None;
+
+    #[no_mangle]
+    extern "C" fn init() {
         unsafe {
-            CATALOG_ADMIN.get_or_insert_with(|| *exec_context.actor_id());
-        }
-        Self { data, exec_context }
-    }
-}
-
-#[gservice]
-impl<'a, TExecContext> Catalog<'a, TExecContext>
-where
-    TExecContext: ExecContext,
-{
-    pub fn add_parts(&mut self, parts: PartMap<PartId, Part>) -> Result<PartMap<PartId, Part>> {
-        self.require_admin()?;
-
-        if parts.is_empty() {
-            return Err(Error::ZeroLengthPassed);
-        }
-
-        for (part_id, part) in &parts {
-            if *part_id == 0 {
-                return Err(Error::PartIdCantBeZero);
-            }
-            if self.data.parts.contains_key(part_id) {
-                return Err(Error::PartAlreadyExists);
-            }
-            self.data.parts.insert(*part_id, part.clone());
-        }
-
-        Ok(parts)
-    }
-
-    pub fn remove_parts(&mut self, part_ids: Vec<PartId>) -> Result<Vec<PartId>> {
-        self.require_admin()?;
-
-        if part_ids.is_empty() {
-            return Err(Error::ZeroLengthPassed);
-        }
-
-        if !part_ids
-            .iter()
-            .all(|part_id| self.data.parts.contains_key(part_id))
-        {
-            return Err(Error::PartDoesNotExist);
-        }
-
-        for part_id in &part_ids {
-            self.data.parts.remove(part_id);
-        }
-
-        Ok(part_ids)
-    }
-
-    pub fn part(&self, part_id: PartId) -> Option<Part> {
-        self.data.parts.get(&part_id).cloned()
-    }
-
-    pub fn add_equippables(
-        &mut self,
-        part_id: PartId,
-        collection_ids: Vec<CollectionId>,
-    ) -> Result<(PartId, Vec<CollectionId>)> {
-        self.require_admin()?;
-
-        if collection_ids.is_empty() {
-            return Err(Error::ZeroLengthPassed);
-        }
-
-        let part = self
-            .data
-            .parts
-            .get_mut(&part_id)
-            .ok_or(Error::PartDoesNotExist)?;
-
-        let equippable = if let Part::Slot(SlotPart { equippable, .. }) = part {
-            equippable
-        } else {
-            return Err(Error::WrongPartFormat);
-        };
-
-        for collection_id in &collection_ids {
-            equippable.push(*collection_id);
-        }
-
-        Ok((part_id, collection_ids))
-    }
-
-    pub fn remove_equippable(
-        &mut self,
-        part_id: PartId,
-        collection_id: CollectionId,
-    ) -> Result<(PartId, CollectionId)> {
-        self.require_admin()?;
-
-        let part = self
-            .data
-            .parts
-            .get_mut(&part_id)
-            .ok_or(Error::PartDoesNotExist)?;
-
-        if let Part::Slot(SlotPart { equippable, .. }) = part {
-            equippable.retain(|x| x != &collection_id);
-        } else {
-            return Err(Error::WrongPartFormat);
-        }
-
-        Ok((part_id, collection_id))
-    }
-
-    pub fn reset_equippables(&mut self, part_id: PartId) -> Result<()> {
-        self.require_admin()?;
-
-        let part = self
-            .data
-            .parts
-            .get_mut(&part_id)
-            .ok_or(Error::PartDoesNotExist)?;
-
-        let equippable = if let Part::Slot(SlotPart { equippable, .. }) = part {
-            equippable
-        } else {
-            return Err(Error::WrongPartFormat);
-        };
-
-        *equippable = vec![];
-        self.data.is_equippable_to_all.retain(|x| x != &part_id);
-
-        Ok(())
-    }
-
-    pub fn set_equippables_to_all(&mut self, part_id: PartId) -> Result<()> {
-        self.require_admin()?;
-
-        let part = self
-            .data
-            .parts
-            .get(&part_id)
-            .ok_or(Error::PartDoesNotExist)?;
-
-        if let Part::Fixed { .. } = part {
-            return Err(Error::WrongPartFormat);
-        }
-
-        self.data.is_equippable_to_all.insert(part_id);
-
-        Ok(())
-    }
-
-    pub fn equippable(&self, part_id: PartId, collection_id: CollectionId) -> Result<bool> {
-        for equippable in &self.data.is_equippable_to_all {
-            if equippable == &part_id {
-                return Ok(true);
-            }
-        }
-
-        let part = self
-            .data
-            .parts
-            .get(&part_id)
-            .ok_or(Error::PartDoesNotExist)?;
-
-        if let Part::Slot(SlotPart { equippable, .. }) = part {
-            if equippable.iter().any(|x| x == &collection_id) {
-                Ok(true)
-            } else {
-                Ok(false)
-            }
-        } else {
-            Err(Error::WrongPartFormat)
+            PROGRAM = Some(Program::new());
         }
     }
 
-    fn require_admin(&self) -> Result<()> {
-        if self.exec_context.actor_id() != catalog_admin() {
-            return Err(Error::NotAllowedToCall);
-        }
-
-        Ok(())
+    #[gstd::async_main] // Make async optional
+    async fn main() {
+        let input_bytes = msg::load_bytes().expect("Failed to read input");
+        let mut catalog = unsafe { PROGRAM.as_ref().unwrap() }.catalog();
+        let output_bytes = requests::process(&mut catalog, &input_bytes).await;
+        msg::reply_bytes(output_bytes, 0).expect("Failed to send output");
     }
-}
-
-fn catalog_admin() -> &'static ActorId {
-    unsafe { CATALOG_ADMIN.as_ref().unwrap() }
 }
