+++ examples/rmrk/catalog/app/src/services/errors.rs
@@ -1,4 +1,4 @@
-use sails_rtl::{Encode, TypeInfo};
+use sails_rtl_gstd::{Encode, TypeInfo};
 
 #[derive(Encode, TypeInfo)]
 pub enum Error {
