+++ examples/rmrk/resource/app/Cargo.toml
@@ -8,6 +8,7 @@ gstd.workspace = true
 parity-scale-codec = { workspace = true, features = ["derive"] }
 sails-macros.workspace = true
 sails-rtl.workspace = true
+sails-rtl-gstd.workspace = true
 sails-sender.workspace = true
 sails-service-meta.workspace = true
 scale-info = { workspace = true, features = ["derive"] }
