+++ examples/rmrk/resource/app/src/lib.rs
@@ -1,207 +1,71 @@
 #![no_std]
 
-pub use catalogs::Client as CatalogClientImpl;
-use catalogs::Service as CatalogClient;
-use errors::{Error, Result};
-use gstd::ActorId as GStdActorId;
-use resources::{ComposedResource, PartId, Resource, ResourceId};
-use sails_macros::gservice;
-use sails_rtl::{collections::HashMap, *};
+use catalogs::Client as CatalogClient;
+use sails_rtl_gstd::GStdExecContext;
+use sails_sender::GStdSender;
+use services::ResourceStorage;
 
 mod catalogs;
-pub mod errors;
-pub mod resources;
-
-static mut RESOURCE_STORAGE_DATA: Option<ResourceStorageData> = None;
-
-static mut RESOURCE_STORAGE_ADMIN: Option<ActorId> = None;
+// Exposed publicly because of service metadata for IDL
+// Will be superseded by program metadata, thus pub won't be needed
+pub mod services;
 
 #[derive(Default)]
-struct ResourceStorageData {
-    resources: HashMap<ResourceId, Resource>,
-}
+pub struct Program;
 
-pub struct ResourceStorage<TExecContext, TCatalogClient> {
-    exec_context: TExecContext,
-    catalog_client: TCatalogClient,
-}
-
-#[gservice]
-impl<TExecContext, TCatalogClient> ResourceStorage<TExecContext, TCatalogClient>
-where
-    TExecContext: ExecContext,
-    TCatalogClient: CatalogClient,
-{
-    pub fn new(exec_context: TExecContext, catalog_client: TCatalogClient) -> Self {
-        unsafe {
-            RESOURCE_STORAGE_DATA.get_or_insert_with(Default::default);
-            RESOURCE_STORAGE_ADMIN.get_or_insert_with(|| *exec_context.actor_id());
-        }
-        Self {
-            exec_context,
-            catalog_client,
-        }
+impl Program {
+    // Initialize program and seed hosted services
+    pub fn new() -> Self {
+        let exec_context = GStdExecContext::default();
+        ResourceStorage::<_, CatalogClient>::seed(exec_context);
+        Self
     }
 
-    pub fn add_resource_entry(
-        &mut self,
-        resource_id: ResourceId,
-        resource: Resource,
-    ) -> Result<(ResourceId, Resource)> {
-        self.require_admin()?;
-
-        if resource_id == 0 {
-            return Err(Error::ZeroResourceId);
-        }
-
-        if resource_storage_data_mut()
-            .resources
-            .insert(resource_id, resource.clone())
-            .is_some()
-        {
-            return Err(Error::ResourceAlreadyExists);
-        }
-
-        Ok((resource_id, resource))
+    // Expose hosted service
+    pub fn resource_storage(&self) -> ResourceStorage<GStdExecContext, CatalogClient> {
+        let exec_context = GStdExecContext::default();
+        let sender = GStdSender::default();
+        ResourceStorage::new(exec_context, CatalogClient::new(sender))
     }
 
-    pub async fn add_part_to_resource(
-        &mut self,
-        resource_id: ResourceId,
-        part_id: PartId,
-    ) -> Result<PartId> {
-        self.require_admin()?;
-
-        let resource = resource_storage_data_mut()
-            .resources
-            .get_mut(&resource_id)
-            .ok_or(Error::ResourceNotFound)?;
-
-        if let Resource::Composed(ComposedResource { base, parts, .. }) = resource {
-            let part_call = self
-                .catalog_client
-                .part(part_id)
-                .send(GStdActorId::from_slice(base.as_ref()).unwrap())
-                .await
-                .unwrap();
-            let part_response = part_call.response().await.unwrap();
-            if part_response.is_none() {
-                return Err(Error::PartNotFound);
-            }
-            parts.push(part_id);
-        } else {
-            return Err(Error::WrongResourceType);
-        }
-
-        Ok(part_id)
+    // Will be generated by the `ginject_defaults` macro from the `new` method
+    fn __new(exec_context: GStdExecContext) -> Self {
+        ResourceStorage::<_, CatalogClient>::seed(exec_context);
+        Self
     }
 
-    pub fn resource(&self, resource_id: ResourceId) -> Result<Resource> {
-        resource_storage_data_mut()
-            .resources
-            .get(&resource_id)
-            .cloned()
-            .ok_or(Error::ResourceNotFound)
-    }
-
-    fn require_admin(&self) -> Result<()> {
-        if self.exec_context.actor_id() != resource_storage_admin() {
-            return Err(Error::NotAuthorized);
-        }
-        Ok(())
+    // Will be generated by the `ginject_defaults` macro from the `resource_storage` method
+    fn __resource_storage(
+        &self,
+        exec_context: GStdExecContext,
+        sender: GStdSender,
+    ) -> ResourceStorage<GStdExecContext, CatalogClient> {
+        ResourceStorage::new(exec_context, CatalogClient::new(sender))
     }
 }
 
-fn resource_storage_data_mut() -> &'static mut ResourceStorageData {
-    unsafe { RESOURCE_STORAGE_DATA.as_mut().unwrap() }
-}
-
-fn resource_storage_admin() -> &'static ActorId {
-    unsafe { RESOURCE_STORAGE_ADMIN.as_ref().unwrap() }
-}
-
-#[cfg(test)]
-mod tests {
+// The below code will be generated by the `gprogram` macro
+#[cfg(target_arch = "wasm32")]
+pub mod wasm {
     use super::*;
-    use catalogs::{Error, Part};
-    use resources::BasicResource;
-    use sails_rtl::{collections::BTreeMap, ActorId, Result as RtlResult};
-    use sails_sender::Call;
+    use sails_rtl_gstd::{gstd, gstd::msg};
+    use services::requests;
 
-    #[test]
-    fn test_add_resource_entry() {
-        let mut resource_storage =
-            ResourceStorage::new(ExecContextMock { actor_id: 1.into() }, CatalogClientMock);
-        let resource = Resource::Basic(BasicResource {
-            src: "src".to_string(),
-            thumb: None,
-            metadata_uri: "metadata_uri".to_string(),
-        });
-        let (actual_resource_id, actual_resource) = resource_storage
-            .add_resource_entry(1, resource.clone())
-            .unwrap();
-        assert_eq!(actual_resource_id, 1);
-        assert_eq!(actual_resource, resource);
-    }
+    static mut PROGRAM: Option<Program> = None;
 
-    struct ExecContextMock {
-        actor_id: ActorId,
-    }
-
-    impl ExecContext for ExecContextMock {
-        fn actor_id(&self) -> &ActorId {
-            &self.actor_id
+    #[no_mangle]
+    extern "C" fn init() {
+        let exec_context = GStdExecContext::default();
+        unsafe {
+            PROGRAM = Some(Program::new());
         }
     }
 
-    struct CatalogClientMock;
-
-    impl CatalogClient for CatalogClientMock {
-        fn add_parts(
-            &mut self,
-            _parts: BTreeMap<u32, Part>,
-        ) -> Call<RtlResult<BTreeMap<u32, Part>, Error>> {
-            unimplemented!()
-        }
-
-        fn remove_parts(&mut self, _part_ids: Vec<u32>) -> Call<RtlResult<Vec<u32>, Error>> {
-            unimplemented!()
-        }
-
-        fn add_equippables(
-            &mut self,
-            _part_id: u32,
-            _collection_ids: Vec<ActorId>,
-        ) -> Call<RtlResult<(u32, Vec<ActorId>), Error>> {
-            unimplemented!()
-        }
-
-        fn remove_equippable(
-            &mut self,
-            _part_id: u32,
-            _collection_id: ActorId,
-        ) -> Call<RtlResult<(u32, ActorId), Error>> {
-            unimplemented!()
-        }
-
-        fn reset_equippables(&mut self, _part_id: u32) -> Call<RtlResult<(), Error>> {
-            unimplemented!()
-        }
-
-        fn set_equippables_to_all(&mut self, _part_id: u32) -> Call<RtlResult<(), Error>> {
-            unimplemented!()
-        }
-
-        fn part(&self, _part_id: u32) -> Call<Option<Part>> {
-            unimplemented!()
-        }
-
-        fn equippable(
-            &self,
-            _part_id: u32,
-            _collection_id: ActorId,
-        ) -> Call<RtlResult<bool, Error>> {
-            unimplemented!()
-        }
+    #[gstd::async_main]
+    async fn main() {
+        let input_bytes = msg::load_bytes().expect("Failed to read input");
+        let mut resource_storage = unsafe { PROGRAM.as_ref().unwrap() }.resource_storage();
+        let output_bytes = requests::process(&mut resource_storage, &input_bytes).await;
+        msg::reply_bytes(output_bytes, 0).expect("Failed to send output");
     }
 }
