+++ examples/rmrk/resource/app/src/services/errors.rs
