+++ examples/rmrk/resource/app/src/services/resources.rs
